+++ source/jormungandr/jormungandr/scenarios/helper_classes/__init__.py
@@ -26,12 +26,12 @@
 # IRC #navitia on freenode
 # https://groups.google.com/d/forum/navitia
 # www.navitia.io
-from PlacesFreeAccess import PlacesFreeAccess
-from StreetNetworkPath import StreetNetworkPathPool
-from FallbackDurations import FallbackDurationsPool
-from PlaceByUri import PlaceByUri
-from PtJourney import PtJourneyPool
-from ProximitiesByCrowfly import ProximitiesByCrowflyPool
-from CompletePtJourney import wait_and_complete_pt_journey
-from Exceptions import PtException, EntryPointException
+from places_free_access import PlacesFreeAccess
+from streetnetwork_path import StreetNetworkPathPool
+from fallback_durations import FallbackDurationsPool
+from place_by_uri import PlaceByUri
+from pt_journey import PtJourneyPool
+from proximities_by_crowfly import ProximitiesByCrowflyPool
+from complete_pt_journey import wait_and_complete_pt_journey
+from helper_exceptions import PtException, EntryPointException
 from helper_utils import get_entry_point_or_raise, check_final_results_or_raise
