+++ source/jormungandr/jormungandr/scenarios/helper_classes/place_by_uri.py
@@ -26,7 +26,7 @@
 # IRC #navitia on freenode
 # https://groups.google.com/d/forum/navitia
 # www.navitia.io
-import Future
+import helper_future
 
 
 class PlaceByUri:
@@ -40,7 +40,7 @@ class PlaceByUri:
         return self._instance.georef.place(self._uri)
 
     def _async_request(self):
-        self._value = Future.create_future(self._do_request)
+        self._value = helper_future.create_future(self._do_request)
 
     def wait_and_get(self):
         return self._value.wait_and_get()
