+++ source/jormungandr/jormungandr/scenarios/helper_classes/places_free_access.py
@@ -26,7 +26,7 @@
 # IRC #navitia on freenode
 # https://groups.google.com/d/forum/navitia
 # www.navitia.io
-import Future
+import helper_future
 from navitiacommon import type_pb2
 from jormungandr import utils
 from collections import namedtuple
@@ -74,7 +74,7 @@ class PlacesFreeAccess:
         return PlaceFreeAccessResult(crowfly, odt)
 
     def _async_request(self):
-        self._value = Future.create_future(self._do_request)
+        self._value = helper_future.create_future(self._do_request)
 
     def wait_and_get(self):
         return self._value.wait_and_get()
