+++ source/jormungandr/jormungandr/scenarios/helper_classes/proximities_by_crowfly.py
@@ -26,7 +26,7 @@
 # IRC #navitia on freenode
 # https://groups.google.com/d/forum/navitia
 # www.navitia.io
-import Future
+import helper_future
 from jormungandr.street_network.street_network import StreetNetworkPathType
 from helper_utils import get_max_fallback_duration
 from jormungandr import utils
@@ -78,7 +78,7 @@ class ProximitiesByCrowfly:
         return []
 
     def _async_request(self):
-        self._value = Future.create_future(self._do_request)
+        self._value = helper_future.create_future(self._do_request)
 
     def wait_and_get(self):
         return self._value.wait_and_get()
